Miscellaneous code cleanup.
This patch consists of various bits of cleanup throughout the code base,
such as improving readability, deleting unnecessary code, and adding
conditionals (BBGE_BUILD_SDL in particular) around system-specific code.

+++ BBGE/ScreenTransition.cpp
@@ -138,8 +138,8 @@ void ScreenTransition::onRender()
 	if (alpha.x == 0) return;
 	
 #ifdef BBGE_BUILD_OPENGL
-	int width2 = width/2;
-	int height2 = height/2;
+	float width2 = float(width)/2;
+	float height2 = float(height)/2;
 	
 	const float pw = float(windowWidth)/float(textureWidth);
 	const float ph = float(windowHeight)/float(textureHeight);
@@ -167,7 +167,7 @@ void ScreenTransition::onRender()
 		glTranslatef(0.5f,0.0f,0.0f);
 	}
 #endif
-	
+
 	glBegin(GL_QUADS);
 		//glNormal3f( 0.0f, 0.0f, 1.0f);
 		//glColor4f(color.x, color.y, color.z, alpha.getValue());
